Bugfix-Spielende
Code Anpassungen
