+++ app/src/main/java/at/aau/serg/websocketdemoapp/networking/StompHandler.java
@@ -26,8 +26,8 @@ public class StompHandler {
     private static StompHandler instance;
     private static final String TAG_NETWORK = "Network";
     private static final String TAG_RECEIVED = "Received";
-    //private static final String actualServerUrl = "ws://unified-officially-snake.ngrok-free.app/websocket-example-broker";
-    private static final String actualServerUrl = "ws://10.0.2.2:8080/websocket-example-broker";
+    private static final String actualServerUrl = "ws://unified-officially-snake.ngrok-free.app/websocket-example-broker";
+    //private static final String actualServerUrl = "ws://10.0.2.2:8080/websocket-example-broker";
 
     public StompHandler() {
         stompClient = Stomp.over(Stomp.ConnectionProvider.OKHTTP, actualServerUrl);
